refactor(frontend): refactor registration and user edition to use a common UserForm component, which reduces code duplication
