+++ frontend/src/app/features/registration/registration.component.html
@@ -1,77 +1,3 @@
 <h1 i18n="registration title|title of the component used to register user@@user.registration.title">Registration</h1>
 
-<form [formGroup]="form" (ngSubmit)="submit()">
-    <mat-form-field>
-        <mat-label i18n="email|email input label@@input.email.label">Email</mat-label>
-        <input matInput formControlName="email" i18n-placeholder="email placeholder|email input placeholder@@input.email.placeholder" placeholder="Enter your email">
-        @if(form.get('email')?.hasError('emailExists')) {
-            <mat-hint><span i18n="email exists|inform the user the email already exists@@info.email_exists">Email already exists</span> <a routerLink="/login">Log in</a></mat-hint>
-        }
-        <app-input-backend-errors [control]="form.get('email')" />
-        <mat-hint class="email-hint" align="end">
-            <app-status-icon [isValid]="!email?.invalid" />
-        </mat-hint>
-    </mat-form-field>
-
-    <mat-form-field>
-        <mat-label i18n="username|username input label@@input.username.label">Username</mat-label>
-        <input matInput minlength="5" formControlName="username" i18n-placeholder="username placeholder|username input placeholder@@input.username.placeholder" placeholder="Type your username">
-        @if(form.get('username')?.hasError('usernameExists')) {
-            <mat-hint><span i18n="username exists|inform the user the username already exists@@info.username_exists">Username already taken</span> <a routerLink="/login">Log in</a></mat-hint>
-        }
-        <app-input-backend-errors [control]="form.get('username')" />
-        <mat-hint class="username-hint" align="end">
-            <app-status-icon [isValid]="!username?.invalid" />
-        </mat-hint>
-    </mat-form-field>
-
-    <mat-form-field>
-        <mat-label i18n="@@input.password.label">Password</mat-label>
-        <input maxlength="128" matInput formControlName="password" type="password" i18n-placeholder="password placeholder@@input.password.placeholder" placeholder="Enter your password">
-        @if(password?.invalid) {
-            @if(password?.errors !== null && password?.errors?.['passwordStrength'] !== null) {
-                @let errors = password!.errors?.['passwordStrength'];
-                @if(errors !== undefined && errors?.message !== undefined) {
-                    <mat-hint class="password-hint" align="start">
-                        {{  errors.message }}
-                    </mat-hint>
-                }
-            }
-        }
-
-        <mat-hint class="password-hint" align="end">
-            <app-status-icon [isValid]="!password?.invalid" />
-        </mat-hint>
-    </mat-form-field>
-
-    <mat-form-field>
-        <mat-label i18n="first name|first name input label@@input.first_name.label">First name</mat-label>
-        <input matInput type="text" formControlName="firstName" i18n-placeholder="first name placeholder|first name input placeholder@@input.first_name.placeholder"  placeholder="Enter your first name">
-        <app-input-backend-errors [control]="form.get('firstName')" />
-        <mat-hint class="last-name-hint" align="end">
-            <app-status-icon [isValid]="!firstName?.invalid"></app-status-icon>
-        </mat-hint>
-    </mat-form-field>
-
-    <mat-form-field>
-        <mat-label i18n="last name|last name input label@@input.last_name.label">Last name</mat-label>
-        <input matInput type="text" formControlName="lastName"  i18n-placeholder="last name placeholder|last name input placeholder@@input.last_name.placeholder" placeholder="Enter your last name">
-        <app-input-backend-errors [control]="form.get('lastName')" />
-        <mat-hint class="last-name-hint" align="end">
-            <app-status-icon [isValid]="!lastName?.invalid"></app-status-icon>
-        </mat-hint>
-    </mat-form-field>
-
-    <input type="hidden" formControlName="lang" name="lang" [value]="lang">
-
-    <button mat-raised-button type="submit" [disabled]="form.invalid || isSubmitting">
-        @if (isSubmitting) { 
-            <span i18n="loading|inform the user loading occurs@@info.loading">Loading...</span>
-        } 
-        @else { 
-            <span i18n="register|Register action label@@action.register">Register</span>
-        }
-    </button>
-</form>
-    
-
+<app-user-form [form]="form" (submitted)="submit()" [submitLabel]="'Register'" [isSubmitting]="isSubmitting" [showPassword]="true">
