+++ frontend/src/app/features/user/user-edit/user-edit.component.html
@@ -1,0 +1,3 @@
+<h1 i18n="edit user info@@user.information.edit">Edit my information</h1>
+
+<app-user-form [form]="form" [submitLabel]="'Save'" (submitted)="submit()" [showPassword]="false" [isSubmitting]="isSubmitting" />
