+++ frontend/src/app/features/user/user-form/user-form.component.ts
@@ -1,19 +1,11 @@
-import { Component, OnDestroy, OnInit } from '@angular/core';
-import { User } from '../../../core/model/user.interface';
-import { FormBuilder, FormGroup, ReactiveFormsModule, Validators } from '@angular/forms';
+import { Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
+import { FormControl, FormGroup, ReactiveFormsModule } from '@angular/forms';
 import { MatFormField, MatHint, MatLabel } from '@angular/material/form-field';
-import { AuthValidators } from '../../../core/services/auth.validators';
 import { MatInput } from '@angular/material/input';
-import { UserService } from '../../../core/services/user.service';
-import { EditUserRequest } from '../../../core/model/edit-user-request.interface';
-import { catchError, throwError } from 'rxjs';
-import { ApiError } from '../../../core/errors/api-error';
-import { NotificationService } from '../../../core/services/notification.service';
-import { BaseChildComponent } from '../../../core/component/base-child.component';
-import { FormErrorService } from '../../../core/services/form-error.service';
 import { StatusIconComponent } from '../../../layout/shared/status-icon/status-icon.component';
 import { InputBackendErrorsComponent } from "../../../layout/shared/input-backend-errors/input-backend-errors.component";
 import { MatButton } from '@angular/material/button';
+import { LocaleService } from '../../../core/services/locale.service';
 
 @Component({
   selector: 'app-user-form',
@@ -21,59 +13,20 @@ import { MatButton } from '@angular/material/button';
   templateUrl: './user-form.component.html',
   styleUrl: './user-form.component.scss'
 })
-export class UserFormComponent extends BaseChildComponent implements OnInit, OnDestroy {
-  protected user!: User;
+export class UserFormComponent implements OnInit {
+  @Input() form!: FormGroup;
+  @Input() isSubmitting = false;
+  @Input() showPassword = false;
+  @Input() submitLabel: 'Register' | 'Save' = 'Save';
+  @Output() submitted = new EventEmitter<void>();
 
-  protected form!: FormGroup;
-  protected isSubmitting = false;
-
-  constructor(
-    private authValidators: AuthValidators,
-    private fb: FormBuilder,
-    private userService: UserService,
-    private notificationService: NotificationService,
-    private formErrorService: FormErrorService
-  ) {
-    super();
+  public currentLang: string;
+  
+  constructor(private localeService: LocaleService) {
+    this.currentLang = this.localeService.currentLocale.toUpperCase();
   }
-  ngOnDestroy(): void {
-    this.formErrorService.cleanup(); // free subscriptions 
-  }
-
   ngOnInit(): void {
-    this.user  = this.data.user!;
-    this.form = this.fb.group({
-      email: [
-        this.user.email, 
-        [
-          Validators.required,
-          Validators.email
-        ],
-        this.authValidators.checkEmailExistsAsync(this.user.email).bind(this.authValidators)
-      ],
-      username: [
-        this.user.username, 
-        [
-          Validators.required,
-          Validators.minLength(5)
-        ],
-        this.authValidators.checkUsernameExistsAsync(this.user.username).bind(this.authValidators)
-      ],
-      firstName: [
-        this.user.firstName, 
-        [
-          Validators.required,
-          Validators.minLength(1)
-        ]
-      ],
-      lastName: [
-        this.user.lastName, 
-        [
-          Validators.required,
-          Validators.minLength(1)
-        ]
-      ],
-    });
+    this.form.addControl('lang', new FormControl(this.currentLang, []));
   }
 
   get email() {
@@ -92,30 +45,15 @@ export class UserFormComponent extends BaseChildComponent implements OnInit, OnD
     return this.form.get('username');
   }
 
+  get password() {
+    return this.form.get('password');
+  }
+
+  get lang() {
+    return this.form.get('lang');
+  }
+
   submit() :void {
-      if(!this.isSubmitting && this.form.valid) {
-
-        const emailChanged: boolean = this.email?.value  !== this.user.email;
-
-        this.isSubmitting = true;
-        this.userService.editUser(this.form.value as EditUserRequest)
-            .pipe(
-              // take(1),
-              catchError(
-                (error: ApiError) => {
-                  this.isSubmitting = false;
-                  
-                  this.formErrorService.setBackendErrors(this.form, error.errors);
-
-
-                  return throwError(() => error);
-                }
-            ))
-            .subscribe(() => {
-                this.isSubmitting = false;
-                this.notificationService.confirmation("Your information has been updated."+(emailChanged ? " Your new email address needs verification ; please check your emails." : ''));
-                this.success();
-            });
-      }
-    }
+      this.submitted.emit();
+  }
 }
